WI #1593 Use dedicated checker classes for diagnostics.
